Mlayer refractor code
